Make most of the code reuseable for future
